Extract logsetting from server_adapters.py
